requirements.txt
tensorflow==2.14.0
numpy==1.25.0
pillow==10.0.0
streamlit==1.26.0
flask==2.3.6
pyngrok==5.2.1
